requirements.txt
librosa
numpy
pandas
soundfile
torchaudio
torch
faiss-cpu
google-generativeai
transformers
audiocraft
tqdm